requirements.txt
dotenv==0.9.9
numpy==2.3.0
opencv-python==4.11.0.86
pillow==11.2.1
python-dotenv==1.1.0
